requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1    # ORM数据库
flask-migrate==4.0.5       # 数据库迁移
pyjwt==2.8.0               # JWT认证
python-dotenv==1.0.0       # 环境变量
cryptography==41.0.7       # 密码加密